requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.2.0
flask==3.0.2
fonttools==4.48.1
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
trio==0.24.0
trio-websocket==0.11.1
typing-extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0